02 release

+++ services/authentication/src/main/java/com/graduationProject/authentication/saga/sagaParticipators/CreateResident.java
@@ -3,18 +3,20 @@ package com.graduationProject.authentication.saga.sagaParticipators;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.graduationProject.authentication.dto.ResidentDto;
 import com.graduationProject.authentication.dto.saga.SagaResidentDto;
 import com.graduationProject.authentication.dto.saga.SagaResponseDto;
 import com.graduationProject.authentication.kafka.KafkaApi;
 import com.graduationProject.authentication.saga.SagaParticipator;
 import com.graduationProject.authentication.service.ResidentService;
 import com.graduationProject.authentication.type.SagaStatus;
+import org.apache.commons.lang3.exception.ExceptionUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import static com.graduationProject.authentication.topic.ResidentTopics.CreateResidentSagaDone;
-import static com.graduationProject.authentication.topic.ResidentTopics.CreateResidentSagaRevert;
+import javax.transaction.Transactional;
+
+import static com.graduationProject.authentication.topic.ResidentTopic.CreateResidentSagaDone;
+import static com.graduationProject.authentication.topic.ResidentTopic.CreateResidentSagaRevert;
 
 
 @Service
@@ -34,17 +36,19 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
     public void transact(SagaResidentDto sagaResidentDto) {
         //if anything goes wrong it should publish to the revert topic
         try {
+            if (sagaResidentDto.getResidentDto().getUsername().equals("fail")) {
+                throw new IllegalStateException(String.format("Could not create Resident with \n ID: %s \n SagaID: %s",
+                        sagaResidentDto.getResidentDto().getId(),
+                        sagaResidentDto.getSagaId())); //this should trigger a revert of the saga.
+            }
             residentService.addResident(sagaResidentDto.getResidentDto());
             SagaResponseDto sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(),
                     SagaStatus.SUCCESS);
-            if (sagaResidentDto.getResidentDto().getUsername().equals("fail")) {
-                throw new Exception(); //this should trigger a revert of the saga.
-            }
             kafkaApi.publish(CreateResidentSagaDone, new ObjectMapper().writeValueAsString(sagaResponseDto));
-            System.out.println("to topic " + CreateResidentSagaDone + " i sent this: " + sagaResponseDto.toString());
         } catch (Exception e) {
             SagaResponseDto sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(),
                     SagaStatus.FAILED);
+            sagaResponseDto.setErrorMessage(ExceptionUtils.getStackTrace(e));
             try {
                 kafkaApi.publish(CreateResidentSagaDone, new ObjectMapper().writeValueAsString(sagaResponseDto));
             } catch (JsonProcessingException ex) {
@@ -55,6 +59,7 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
 
     }
 
+    @Transactional
     @Override
     public void revert(SagaResidentDto sagaResidentDto) {
         /*
@@ -67,17 +72,22 @@ public class CreateResident implements SagaParticipator<SagaResidentDto> {
          */
         try {
             if (sagaResidentDto.getResidentDto().getPassword().equals("fail")) {
-                throw new Exception();
+                throw new IllegalStateException(String.format("Could not revert creation of Resident with \n ID: %s \n SagaID: %s",
+                        sagaResidentDto.getResidentDto().getId(),
+                        sagaResidentDto.getSagaId()));
             }
-
-            residentService.deleteResident(sagaResidentDto.getResidentDto().getId());
+            if (residentService.residentExists(sagaResidentDto.getResidentDto().getId())) {
+                residentService.deleteResident(sagaResidentDto.getResidentDto().getId());
+            }
             kafkaApi.publish(CreateResidentSagaRevert, new ObjectMapper()
                     .writeValueAsString(new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.SUCCESS)));
         } catch (Exception e) {
             e.printStackTrace();
             try {
+                SagaResponseDto sagaResponseDto = new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.FAILED);
+                sagaResponseDto.setErrorMessage(ExceptionUtils.getStackTrace(e));
                 kafkaApi.publish(CreateResidentSagaRevert, new ObjectMapper()
-                        .writeValueAsString(new SagaResponseDto(sagaResidentDto.getSagaId(), SagaStatus.FAILED)));
+                        .writeValueAsString(sagaResponseDto));
             } catch (JsonProcessingException ex) {
                 ex.printStackTrace();
             }
